+++ src/components/feedback/feedback-step/buy-step/index.tsx
@@ -1,13 +1,23 @@
-import { Button, Flex, Text } from "@chakra-ui/react";
+import { Button, Flex, Input, Text } from "@chakra-ui/react";
 import { useState } from "react";
 import { RiCupFill } from "react-icons/ri";
 import CofiInput from "./coffee-input";
 
 const BuyAKofiStep = () => {
   const [isClicked, setIsClicked] = useState(false);
-
+  const [amount, setAmount] = useState(1);
+  const price = 0.005;
   const handleClick = () => {
     setIsClicked((prev) => !prev);
+  };
+
+  const handleIncreass = () => {
+    setAmount((prev) => prev + 1);
+  };
+
+  const handleDecreass = () => {
+    if (amount === 1) return;
+    setAmount((prev) => prev - 1);
   };
 
   return (
@@ -30,11 +40,22 @@ const BuyAKofiStep = () => {
             <Flex alignItems={"center"}>
               <RiCupFill color={"#FEA55F"} size={"40px"} />
               <Text color={"lightGray"} fontWeight={"bold"} ml={"10px"}>
-                0.005 eth
+                {price} eth
               </Text>
             </Flex>
-            <CofiInput />
+            <CofiInput
+              amount={amount}
+              increass={handleIncreass}
+              decreass={handleDecreass}
+            />
           </Flex>
+          <Input
+            mt={"10px"}
+            disabled
+            value={`${price * amount} eth`}
+            color={"white"}
+            borderRadius={"20px"}
+          />
         </Flex>
       ) : (
         <Button
